requirements.txt
opencv-python>=4.5.0
numpy>=1.21.0
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
scikit-learn>=1.0.0
scipy>=1.7.0
ultralytics>=8.0.0
deep-sort-realtime>=1.3.0 